+++ src/components/b_dropdown.rs
@@ -1,31 +1,47 @@
 use leptos::html::Div;
 use leptos::*;
 
-use crate::MouseEventFn;
+use crate::elements::BButton;
 
 #[component]
 pub fn BDropdown<F, IV>(
     children: Children,
-    #[prop(default = "")] class: &'static str,
+    #[prop(optional, into)] class: TextProp,
+    #[prop(optional, into)] is_right: MaybeSignal<bool>,
+    #[prop(optional, into)] is_up: MaybeSignal<bool>,
+    #[prop(optional, into)] is_hoverable: MaybeSignal<bool>,
     trigger: F,
 ) -> impl IntoView
 where
-    F: Fn() -> IV,
+    F: Fn() -> IV + 'static,
     IV: IntoView,
 {
     let node_ref = create_node_ref::<Div>();
-    let (is_active, set_is_active) = create_signal(false);
+    let is_active = create_rw_signal(false);
 
     let _ = leptos_use::on_click_outside(node_ref, move |_| {
-        set_is_active.set(false);
+        is_active.set(false);
     });
 
+    let toggle_dropdown = move |_| {
+        if is_hoverable.get() {
+            is_active.set(false);
+        } else {
+            is_active.update(|value| *value = !*value);
+        }
+    };
+
     view! {
-        <div node_ref=node_ref class=format!("dropdown {}", class) class:is-active=is_active>
+        <div
+            node_ref=node_ref
+            class=format!("dropdown {}", class.get())
+            class:is-active=is_active
+            class:is-hoverable=is_hoverable
+            class:is-right=is_right
+            class:is-up=is_up
+        >
             <div class="dropdown-trigger">
-                <button class="button" on:click=move |_| { set_is_active.update(|v| *v = !*v) }>
-                    {trigger()}
-                </button>
+                <BButton on:click=toggle_dropdown>{trigger()}</BButton>
             </div>
 
             <div class="dropdown-menu">
@@ -43,16 +59,14 @@ pub fn BDropdownDivider() -> impl IntoView {
 #[component]
 pub fn BDropdownItem(
     children: Children,
-    #[prop(default = "")] class: &'static str,
+    #[prop(optional, into)] class: TextProp,
     #[prop(optional)] href: Option<&'static str>,
-    #[prop(optional, into)] on_click: Option<MouseEventFn>,
 ) -> impl IntoView
 where
 {
     view! {
-        <a class=format!("dropdown-item {}", class) href=href>
+        <a class=format!("dropdown-item {}", class.get()) href=href>
             {children()}
         </a>
     }
-    .optional_event(ev::click, on_click.map(MouseEventFn::into_inner))
 }